Modified Files:  	Buffercache.cc Buffercache.h Client.cc Client.h
Fixed bugs found by Feng


git-svn-id: https://ceph.svn.sf.net/svnroot/ceph@468 29311d96-e01e-0410-9327-a35deaab8ce9

+++ ceph/client/Buffercache.h
@@ -131,7 +131,7 @@ class Dirtybuffers {
   void insert(Bufferhead* bh);
   bool empty() { return _dbufs.empty(); }
   bool exist(Bufferhead* bh);
-  void get_expired(time_t ttl, size_t left_dirty, set<Bufferhead*>& to_flush);
+  void get_expired(time_t ttl, int left_dirty, set<Bufferhead*>& to_flush);
   time_t get_age() { 
     time_t age;
     if (_dbufs.empty()) {
